fix: rename tenant delete package to deletecmd for consistency
- Rename go/cmd/inventario/tenants/delete to go/cmd/inventario/tenants/deletecmd
- Rename delete.go to deletecmd.go within the package
- Update package declarations from 'delete' to 'deletecmd'
- Update imports in tenants.go to use deletecmd package
- Remove temporary update_clean.go file

This ensures consistency with users/deletecmd naming convention and
avoids conflicts with Go's built-in 'delete' function name.

Both tenant and user delete commands now use deletecmd package name
while maintaining 'delete' as the actual command name.
